requirements.txt
pyjwt
boto3
fastapi
uvicorn
python-dotenv
pymongo
python-multipart
jinja2
pyotp